Partial evaluation implemented for search.
It is now possible to provide partial input to SimplePolynomialFunction by calling resolve. This results in another SimplePolynomialFunction, which is capable of evaluating the remaining bits.

Also updated is the test driver, with better support for running automated tests that push test metrics to a graphite instance for measuring perf.